requirements.txt
annotated-types==0.7.0
anyio==4.7.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.1.7
colorama==0.4.6
fastapi==0.115.6
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
packaging==25.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.3
pydantic_core==2.27.1
PyMySQL==1.1.1
pyparsing==3.2.3
python-dotenv==1.0.1
python-multipart==0.0.19
requests==2.32.4
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.32.1
